Matching focus on mobile and desktop, removing component
